new metric for read-only transactions. fixed default aborts of read-only transactions

+++ arangosh/Benchmark/testcases/VersionTest.h
@@ -44,7 +44,7 @@ struct VersionTest : public Benchmark<VersionTest> {
 
   void tearDown() override {}
 
-  void buildRequest(int threadNumber, size_t threadCounter,
+  void buildRequest(size_t threadNumber, size_t threadCounter,
                     size_t globalCounter, BenchmarkOperation::RequestData& requestData) const override {
     requestData.url = _url;
     requestData.type = rest::RequestType::GET;
